+++ src/Controllers/Admin/DashboardController.php
@@ -3,8 +3,9 @@
 namespace Adam\Superauth\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-
-
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Adam\Superauth\Models\Role;
 class DashboardController extends Controller
 {
     /**
@@ -14,7 +15,7 @@ class DashboardController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('moderators');
+        //$this->middleware('moderators')->except('updateRoles');
     }
 
     /**
@@ -24,6 +25,24 @@ class DashboardController extends Controller
      */
     public function index()
     {
-        return view()->first(['admin.dashboard', 'Superauth::admin.dashboard']);
+        $rolse = Role::all();
+        return view()->first(['admin.dashboard', 'Superauth::admin.dashboard'])
+            ->with('roles', $rolse);
+    }
+
+    /**
+     * Update user Roles
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return redirect back
+     */
+    public function updateRoles(Request $request){
+        if (!$request->role) {
+            return redirect()->back()
+                ->with('danger', trans('Superauth::auth.errorAtLeastOneRoleHasToBeSelected'));
+        }
+        Auth::user()->roles()->sync($request->role);
+        return redirect()->back()
+            ->with('success', trans('Superauth::auth.rolesUpdateSuccess'));
     }
 }