+++ src/Controllers/ProfileController.php
@@ -3,7 +3,7 @@
 namespace Adam\Superauth\Controllers;
 
 use App\Http\Controllers\Controller;
-
+use Adam\Superauth\Models\Role;
 class ProfileController extends Controller
 {
     /**
@@ -23,6 +23,8 @@ class ProfileController extends Controller
      */
     public function index()
     {
-        return view()->first(['profile', 'Superauth::profile']);
+        $rolse = Role::all();
+        return view()->first(['profile', 'Superauth::profile'])
+            ->with('roles', $rolse);
     }
 }