+++ src/resources/views/admin/dashboard.blade.php
@@ -1,4 +1,4 @@
-@extends('Superauth::layouts.app')
+@extends('Superauth::admin.layouts.app')
 
 @section('content')
 <div class="container">
@@ -6,14 +6,33 @@
         <div class="col-md-8">
             <div class="card card-default">
                 <div class="card-header">{{trans('Superauth::auth.adminDashboard')}}</div>
-
                 <div class="card-body">
-                    @if (session('status'))
-                        <div class="alert alert-success">
-                            {{ session('status') }}
+                    <h5>{{ trans('Superauth::auth.welcomeToYour') }}
+                        {{ trans('Superauth::auth.dashboard') }},
+                        {{ trans('Superauth::auth.testUpdatingYourRoles') }}</h5>
+                    @include('Superauth::layouts.partials.alerts')
+                    <form method="POST" action="{{ route('roles.update') }}">
+                        @csrf
+                        @foreach($roles as $role)
+                            <div class="form-check">
+                                <input class="form-check-input" type="checkbox" id="inlineCheckbox{{ $role->id }}" name='role[]'
+                                       value="{{ $role->id }}"{{ (Auth::user()->hasRole($role->id))? ' checked': '' }}>
+                                <label class="form-check-label" for="inlineCheckbox{{ $role->id }}">
+                                    {{ $role->name }} | {!!  ($role->moderatorRole())
+                                        ? '(<strong>'.trans('Superauth::auth.moderator').'</strong>)'
+                                        : '(<strong>'.trans('Superauth::auth.user').'</strong>)'  !!}
+                                </label>
+                            </div>
+                        @endforeach
+                        <hr>
+                        <div class="form-group row mb-0">
+                            <div class="col-md-8">
+                                <button type="submit" class="btn btn-primary">
+                                    {{ trans('Superauth::auth.update') }}
+                                </button>
+                            </div>
                         </div>
-                    @endif
-                    {{ trans('Superauth::auth.youAreLoggedIn') }}
+                    </form>
                 </div>
             </div>
         </div>