+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to `Superauth` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - YYYY-MM-DD
+## NEXT - 2018-03-04
 
 ### Added
-- Nothing
+- Testing User roles update at /profile , /admin 
 
 ### Deprecated
 - Nothing